requirements.txt
Flask
requests
Pillow
numpy
tensorflow==2.15.*
python-dotenv
cryptography
flask_oauthlib==0.9.5
matplotlib
scikit-learn==1.1.2
pandas
adversarial-robustness-toolbox
torch
torchvision